+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add time entry autocomplete to edit modal
+
 ## [1.5.0] - 2022-06-01
 
 ### Added